Se modifica el tipo de dato del id pasa de int a Long de Pais

+++ src/main/java/com/djaa/ProyectoCRUDBackend/model/Pais.java
@@ -12,13 +12,14 @@ public class Pais {
 	//Se generan los atributos de pais
 	@Id
 	@GeneratedValue (strategy = GenerationType.AUTO)
-	private int id;
+	private Long id;
 	
 	private String nombre;
 	
 	public Pais() {
 		
 	}
+	
 	//Se crea el constructor
 	public Pais(String nombre) {
 		super();
@@ -31,12 +32,10 @@ public class Pais {
 	}
 	
 	
-
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
